Finished implementing the mysql bindings
Changed the name and role of some mysql-related functions so it can
correctly performs query to the Database.

Also added the documentation for all functions in the source file. I
know that it should not be there, but i'll move to their right place
once I am done with the programming of the plugin. promise