Structure modification of the source file
The prompt_password function has been renamed _propt_password and has
been moved to the beginning of the source file. It is now an internal
function of the auth source file.

The name of functions calling the database are now called thanks to the
DB_instance structure, that will be set according to the argument in the
plugin configuration file and the argument passed by mosquitto.