Completed all functions, and cleaned the code
I almost finished programming the plugin, there is still a comment to
add there and there, especially on the plugin.c source file and the
auth.c source file.

There is still one thing to do : define the structure of the config
file that will be used by the plugin to configure itself.

Once this is done, I'll create a test bench and a sample DB to check
that everything is working properly.