First working, but undocumented version
The program is now able to authenticate a client that connects to the
broker with a username/password couple using a mysql database and the argon2
hashing algorithm.

I corrected a few bugs concerning the mysql bindings that resulted into
"out of sync" errors . This was due to some missing <mysql_stmt_store_results>
calls in the different mysql-related functions.

I also corrected a bit the argon2 bindings in the crypto.c source file
because of the way the function writes the hash in the given char
buffer.

XXX :

	I erased the different PSK-related code because I want to
	change the way that this will be working. The first that I want to merge
	on master will be a version that will only be able to authenticate a
	client using the username/password field of the connect message of the
	mqtt protocol. In further versions, I'd like to add support for
	PSK-based authentication.

TODO :
	Document all functions, and remove useless calls to log that
	were used for the debug.

	Update the README according to the changes made for the first
	version.

	Create a small tool that can generate a new line in the DB to
	create new clients.